+++ _setuputils.py
@@ -160,9 +160,6 @@ def _system(cmd, cwd=None, env=None, **kwargs):
     if 'shell' not in kwargs:
         kwargs['shell'] = True
     env = env or os.environ.copy()
-    if env.get('CC', '').startswith('ccache '):
-        # Running configure scripts under ccache just adds overhead.
-        env['CC'] = env['CC'][7:]
     return check_call(cmd, cwd=cwd, env=env, **kwargs)