+++ src/gevent/_socket2.py
@@ -6,6 +6,7 @@ from __future__ import absolute_import
 
 # Our import magic sadly makes this warning useless
 # pylint: disable=undefined-variable
+import sys
 
 from gevent import _socketcommon
 from gevent._util import copy_globals